+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added `test.sh` for a "one-click" verification of the client-server setup.
 - Renamed `test-project` to `example-project` and included it in the repository to serve as a ready-to-use test case.
 
 ## [0.4.0] - 2026-02-09